gateway/requirements.txt
fastapi~=0.61.1
uvicorn~=0.12.1
requests~=2.24.0
python-dotenv~=0.14.0